% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep_calcs.R
\name{calculate_epa}
\alias{calculate_epa}
\title{Add EPA calculations
This is only for D1 football}
\usage{
calculate_epa(
  clean_pbp_dat,
  ep_model = cfbscrapR:::ep_model,
  fg_model = cfbscrapR:::fg_model
)
}
\arguments{
\item{clean_pbp_dat}{Clean PBP DataFrame (as pulled from cfb_pbp_dat)}

\item{ep_model}{EP Model}

\item{fg_model}{FG Model}
}
\description{
Extracts raw game by game data.
}
\examples{

}
\keyword{internal}
