% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_betting_line.R
\name{cfb_betting_line}
\alias{cfb_betting_line}
\title{Get betting line information}
\usage{
cfb_betting_line(year, week = 1, team = NULL, conference = NULL,
  season_type = "regular")
}
\arguments{
\item{year}{Select year, (example: 2018)}

\item{week}{Select week, this is optional (also numeric)}

\item{team}{Select team name (example: Texas, Texas A&M, Clemson)}

\item{conference}{Conference abbreviation filter}

\item{season_type}{Select Season Type (regular,postseason,both)}
}
\description{
Get betting line information
}
\examples{

}
\keyword{internal}
