% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_conf_team.R
\name{cfb_conf_team}
\alias{cfb_conf_team}
\title{Lists all teams in conference
Current support only for D1}
\usage{
cfb_conf_team(conference)
}
\arguments{
\item{conference}{Conference (Select a valid conference, D1 football)}
}
\description{
Lists all teams in conference
Current support only for D1
}
\examples{

}
\keyword{internal}
