% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_game_info.R
\name{cfb_game_info}
\alias{cfb_game_info}
\title{Get information from games}
\usage{
cfb_game_info(
  year,
  season_type = "regular",
  week = NULL,
  team = NULL,
  conference = NULL
)
}
\arguments{
\item{year}{Year}

\item{season_type}{Select Season Type (regular,postseason,both)}

\item{week}{Week (optional, keep this numeric)}

\item{team}{Team (optional, D1 Team)}

\item{conference}{Conference (option, select an appropriate conference)}
}
\description{
1 - Game info for all games that year
2 - Game infos for a specific week of games in a year
3 - Game info for one team that year (or a specific week)
4 - Game info for specific conference that year (or a specific week)
}
\examples{

cfb_game_info(2018,week=1)

cfb_game_info(2018,conference='SEC')

cfb_game_info(2018,team="Texas A&M")

}
\keyword{Game}
\keyword{Info}
