% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_pbp_data.R
\name{cfb_pbp_data}
\alias{cfb_pbp_data}
\title{Extract CFB (D1) Play by Play Data - For both plays and drives}
\usage{
cfb_pbp_data(
  year,
  season_type = "regular",
  week = 1,
  team = NULL,
  play_type = NULL,
  drive = NULL,
  epa_wpa = FALSE
)
}
\arguments{
\item{year}{Select year, (example: 2018)}

\item{season_type}{Select Season Type (regular,postseason,both)}

\item{week}{Select week, this is optional (also numeric)}

\item{team}{Select team name (example: Texas, Texas A&M, Clemson)}

\item{play_type}{Select play type (example: see the cfb_play_type_df)}

\item{drive}{Enter anything, and you will get general drive infoo}
}
\description{
Extract CFB (D1) Play by Play Data - For both plays and drives
}
\details{
Extracts raw game by game data. Data comes from https://api.collegefootballdata.com
}
\examples{

cfb_pbp_data(year=2019,week=9,team='Texas A&M')


cfb_pbp_data(year=2019,week=9,team='Texas A&M',epa_wpa=T)
}
\keyword{Play-by-Play}
