% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_play_stats_player.R
\name{cfb_play_stats_player}
\alias{cfb_play_stats_player}
\title{Gets player info associated by play}
\usage{
cfb_play_stats_player(gameId)
}
\arguments{
\item{gameId}{Game ID

#' @keywords internal}
}
\description{
Information describes the players involved in the play
this includes passer, receiver, defensive players who
create sacks or picks, etc
}
\details{
gameId can be extracted from cfb_game_info
}
\examples{

cfb_play_stats_player(401110723)
}
