% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_position_recruiting.R
\name{cfb_position_recruiting}
\alias{cfb_position_recruiting}
\title{CFB Recruiting Information - Position Groups}
\usage{
cfb_position_recruiting(
  start_year = NULL,
  end_year = NULL,
  team = NULL,
  conference = NULL
)
}
\arguments{
\item{start_year}{Year (*Integer*, 4 digits YYYY, optional)}

\item{end_year}{End of time period - 2020 is max value currently (*Integer* 4 digits YYYY, optional)}

\item{team}{Team information (*String* optional)}

\item{conference}{Conference to filter on, abbreviated (*String* optional)}
}
\description{
If only start_year is provided, function will get cfb recruiting information based
on position groups during that year for all FBS teams.
}
\examples{

cfb_position_recruiting(2018)

cfb_position_recruiting(2018, team="Texas")

cfb_position_recruiting(2016, 2020, team="Virginia")

cfb_position_recruiting(2015, 2020, conference = "SEC")

}
