% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_rankings.R
\name{cfb_rankings}
\alias{cfb_rankings}
\title{Gets CFB rankings at a specific week}
\usage{
cfb_rankings(year, week)
}
\arguments{
\item{year}{Year}

\item{week}{Week

#' @keywords internal}
}
\description{
Postseason polls are after Week 13
}
\examples{

cfb_play_stats_player(401110723)
}
