% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_srs_ranking.R
\name{cfb_srs_ranking}
\alias{cfb_srs_ranking}
\title{Get SRS historical rating data}
\usage{
cfb_srs_ranking(year = NULL, team = NULL, conference = NULL)
}
\arguments{
\item{year}{Year (optional)}

\item{team}{Team information (optional)}

\item{confefence}{Conference Abbreivation - S&P+ information by confderence}
}
\description{
Get SRS historical rating data
}
\examples{

cfb_srs_ranking(year=2019)

cfb_srs_ranking(team='Texas A&M')

cfb_srs_ranking(2019,team="Texas")


cfb_srs_ranking(year=2018,conference='SEC')
}
\keyword{SRS}
