% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_team.R
\name{cfb_team}
\alias{cfb_team}
\title{Lists all teams in conference or all FBS
Current support only for D1}
\usage{
cfb_team(conference = NULL, fbs_all = T, year = NULL)
}
\arguments{
\item{conference}{Conference (Select a valid conference, D1 football)}
}
\description{
Lists all teams in conference or all FBS
Current support only for D1
}
\examples{


cfb_team(conference="SEC")
}
\keyword{Teams}
