% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_team_roster_extract.R
\name{cfb_team_roster_data}
\alias{cfb_team_roster_data}
\title{Get a teams full roster}
\usage{
cfb_team_roster_data(team)
}
\arguments{
\item{team}{Team (Select a valid team, D1 football)}
}
\description{
Extracts raw game by game data.
}
\examples{

}
\keyword{internal}
