% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_team_talent.R
\name{cfb_team_talent}
\alias{cfb_team_talent}
\title{Get composite team talent rankings for all teams in a given year or
all rankings from 2015 - 2018}
\usage{
cfb_team_talent(year = NULL)
}
\arguments{
\item{year}{Year}
}
\description{
Extracts team talent rankings as sourced from 247 rankings
}
\examples{


cfb_team_talent()

cfb_team_talent(year=2018)
}
\keyword{internal}
