% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_pbp.R
\name{clean_pbp_dat}
\alias{clean_pbp_dat}
\title{Clean PBP
This is only for D1 football}
\usage{
clean_pbp_dat(raw_df)
}
\arguments{
\item{clean_pbp_dat}{PBP DataFrame (as pulled from cfb_pbp_dat)}
}
\description{
Extracts raw game by game data.
}
\examples{

}
\keyword{internal}
