% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp_calcs.R
\name{create_wpa}
\alias{create_wpa}
\title{Add WPA calculations
This is only for D1 football}
\usage{
create_wpa(df, wp_model = cfbscrapR:::wp_model)
}
\arguments{
\item{df}{Clean PBP DF with EPA calculations}

\item{wp_model}{WPA Model}
}
\description{
Extracts raw game by game data.
}
\examples{

}
\keyword{internal}
