% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn_getFPI.R
\name{espn_getFPI}
\alias{espn_getFPI}
\title{Get the ESPN Football Power Index}
\usage{
espn_getFPI(yr = format(Sys.Date(), "\%Y"), wk = NULL)
}
\arguments{
\item{yr}{Year of the requested rankings}

\item{wk}{Week of season through which the index should be requested}
}
\value{
A data frame of the requested leaderboard
}
\description{
Extracts an NCAAF Football Power Index ranking from ESPN.com and
returns it as a data frame.
}
\examples{
# Capture Week 2 FPI rankings
espn_getFPI(2015, wk = 2)
}
