% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wpa.R
\name{plot_wpa}
\alias{plot_wpa}
\title{Plot the WPA for a specific game}
\usage{
plot_wpa(dat, game_id = NULL, away_color, home_color, winner = "home")
}
\arguments{
\item{df}{Play-by-Play data frame (can be retrieved from cfb_pbp_data)}
}
\description{
Plot the WPA for a specific game
}
\examples{



df = cfb_pbp_data(year=2019,week=9,team='Texas A&M',epa_wpa=T)

plot_wpa(df,away_color="",home_color="",)

}
\keyword{internal}
