% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn_getFPI.R
\name{weekKey}
\alias{weekKey}
\title{Scrape key attribute for FPI week}
\usage{
weekKey(wk, url)
}
\value{
Character string containing url key
}
\description{
Scrape key attribute for FPI week
}
\keyword{internal}
