% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_game_info.R
\name{cfb_game_info}
\alias{cfb_game_info}
\title{Get information from games (depending on a bunch of parameters)
1 - Game info for all games that year
2 - Game infos for a specific week of games in a year
3 - Game info for one team that year (or a specific week)
4 - Game info for specific conference that year (or a specific week)}
\usage{
cfb_game_info(year, week = NULL, team = NULL, conference = NULL)
}
\arguments{
\item{year}{Year}

\item{week}{Week (optional, keep this numeric)}

\item{team}{Team (optional, D1 Team)}

\item{conference}{Conference (option, select an appropriate conference)}
}
\description{
Extracts team talent rankings as sourced from 247 rankings
}
\examples{

cfb_game_info(2018,week=1)

cfb_game_info(2018,conference='SEC')

cfb_game_info(2018,team="Texas A&M")

}
\keyword{internal}
