% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_pbp_data.R
\name{cfb_pbp_data}
\alias{cfb_pbp_data}
\title{Extract CFB Play by Play Data - For both plays and drives
This is only for D1 football
Data comes from https://api.collegefootballdata.com}
\usage{
cfb_pbp_data(year, week = 1, team = NULL, play_type = NULL,
  drive = NULL)
}
\arguments{
\item{year}{Select year, (example: 2018)}

\item{week}{Select week, this is optional (also numeric)}

\item{team}{Select team name (example: Texas, Texas A&M, Clemson)}

\item{play_type}{Select play type (example: see the cfb_play_type_df)}

\item{drive}{Enter anything, and you will get general drive infoo}
}
\description{
Extracts raw game by game data.
}
\examples{

}
\keyword{internal}
