% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_rankings.R
\name{cfb_rankings}
\alias{cfb_rankings}
\title{Gets Historical CFB poll rankings at a specific week}
\usage{
cfb_rankings(year, week)
}
\arguments{
\item{year}{Year}

\item{week}{Week}
}
\description{
Postseason polls are after Week 13
}
\examples{

cfb_rankings(2018,1)

cfb_rankings(2018,14)
}
\keyword{CFB}
\keyword{Rankings}
